Migrate formatDate and formatDateFromIso over into utils/formatters
