+++ src/components/Office/BillableWeight/ShipmentCard/ShipmentCard.test.jsx
@@ -3,8 +3,8 @@ import { render, screen } from '@testing-library/react';
 
 import ShipmentCard from './ShipmentCard';
 
-import { formatAddressShort, formatDateFromIso } from 'shared/formatters';
-import { formatWeight } from 'utils/formatters';
+import { formatAddressShort } from 'shared/formatters';
+import { formatWeight, formatDateFromIso } from 'utils/formatters';
 import { SHIPMENT_OPTIONS } from 'shared/constants';
 
 const tomorrow = new Date();
