+++ src/components/Office/PaymentRequestCard/PaymentRequestCard.jsx
@@ -11,8 +11,8 @@ import styles from './PaymentRequestCard.module.scss';
 import { HistoryShape } from 'types/router';
 import { PaymentRequestShape } from 'types';
 import { LOA_TYPE, PAYMENT_REQUEST_STATUS } from 'shared/constants';
-import { formatDateFromIso, formatCents } from 'shared/formatters';
-import { toDollarString } from 'utils/formatters';
+import { formatCents } from 'shared/formatters';
+import { toDollarString, formatDateFromIso } from 'utils/formatters';
 import PaymentRequestDetails from 'components/Office/PaymentRequestDetails/PaymentRequestDetails';
 import ConnectedAcountingCodesModal from 'components/Office/AccountingCodesModal/AccountingCodesModal';
 import { groupByShipment } from 'utils/serviceItems';
