+++ src/components/Office/ReviewServiceItems/PaymentReviewed.jsx
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types';
 
 import styles from './PaymentReviewed.module.scss';
 
-import { formatDateFromIso } from 'shared/formatters';
-import { toDollarString } from 'utils/formatters';
+import { toDollarString, formatDateFromIso } from 'utils/formatters';
 
 /**
  * This component represents a section shown in the ReviewDetailsCard at the end of navigation.
