+++ src/scenes/Office/OrdersPanel.jsx
@@ -7,7 +7,7 @@ import { withContext } from 'shared/AppContext';
 import { calculateEntitlementsForMove } from 'shared/Entities/modules/moves';
 import { selectServiceMemberForOrders, updateServiceMember } from 'shared/Entities/modules/serviceMembers';
 import { selectOrdersForMove, updateOrders } from 'shared/Entities/modules/orders';
-import { formatDate } from 'shared/formatters';
+import { formatDate } from 'utils/formatters';
 import { PanelSwaggerField, PanelField, SwaggerValue, editablePanelify } from 'shared/EditablePanel';
 import { createModifiedSchemaForOrdersTypesFlag } from 'utils/featureFlags';
 import { openLinkInNewWindow } from 'shared/utils';
