+++ src/scenes/Office/OrdersViewerPanel.js
@@ -6,7 +6,8 @@ import { Field, FormSection, getFormValues, reduxForm } from 'redux-form';
 import { selectMove } from 'shared/Entities/modules/moves';
 import { selectServiceMemberForOrders, updateServiceMember } from 'shared/Entities/modules/serviceMembers';
 import { selectOrdersForMove, updateOrders } from 'shared/Entities/modules/orders';
-import { formatDate, formatDateTime } from 'shared/formatters';
+import { formatDateTime } from 'shared/formatters';
+import { formatDate } from 'utils/formatters';
 import { editablePanelify, PanelField, PanelSwaggerField, RowBasedHeader } from 'shared/EditablePanel';
 import { SwaggerField } from 'shared/JsonSchemaForm/JsonSchemaField';
 import DutyLocationSearchBox from 'components/DutyLocationSearchBox/DutyLocationSearchBox';
