+++ src/shared/EditablePanel/index.jsx
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { get } from 'lodash';
 
-import { formatCents, formatDate } from 'shared/formatters';
+import { formatCents } from 'shared/formatters';
+import { formatDate } from 'utils/formatters';
 import Alert from 'shared/Alert';
 
 import './index.css';
